+++ changelog.md
@@ -4,16 +4,19 @@
 
 - **New:**
   - Allow about dialog to be resized
+  - Add [#TESTSTATUS](docs/commands/commands.md#teststatus) to send commands to test the status display bar
+  - Add [#TESTSCREEN](docs/commands/commands.md#testscreen) to display basic display data
 - **Fixed:**
   - Code editor: Fix find text context menu
   - Fix resizing command input when status or button bar are hidden/shown
   - Character manger: Do not check for map database when trying to import character database
 - **Changed:**
-  - Update electron 34.2.0 to 36.3.1
-  - Update mathjs 14.2.0 to 14.5.1
+  - Update electron 34.2.0 to 36.3.2
+  - Update mathjs 14.2.0 to 14.5.2
   - Update electron-updater 6.6.0 to 6.6.2
   - Update better-sqlite3 11.8.1 to 11.9.1
   - Update yargs-parser 21.1.1 to 22.0.0
+  - Update ace-editor 1.35.0 to 1.41.0
 
 ## 1.6.1 2025-02-23